requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
bcrypt>=4.0.0